pyproject.toml
[project]
name = "toastformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "hydra-core==1.3.2",
    "matplotlib==3.9.1",
    "scikit-learn==1.5.2",
    "setuptools==72.1.0",
    "torch==2.4",
    "zetascale==2.7.3",
]
